[feature]todoModel层建立 需完善异常处理，目测异常处理没有闭环

+++ module_todo/src/main/java/com/cyxbsmobile_single/module_todo/model/bean/DelPushWrapper.kt
@@ -9,7 +9,7 @@ import java.io.Serializable
  */
 data class DelPushWrapper(
     @SerializedName("")
-    val delTodoList: List<Int>,
+    val delTodoList: List<Long>,
     @SerializedName("sync_time")
     val syncTime: Long,
     @SerializedName("force")
