+++ module_todo/src/main/java/com/cyxbsmobile_single/module_todo/model/bean/TodoListGetWrapper.kt
@@ -13,5 +13,5 @@ data class TodoListGetWrapper(
     @SerializedName("sync_time")
     val syncTime: Long,
     @SerializedName("del_todo_array")
-    var delTodoArray: List<Int> = emptyList()
+    var delTodoArray: List<Long> = emptyList()
 ): Serializable
